Include all platform binaries
This change to the props file will make the native binaries for all platforms
be copied into the NativeBinaries folder regardless of which platform you are
building on.

The new LibGit2Sharp.dll.config will ensure that mono can find the linux and
osx binaries when the are in the NativeBinaries folder.