Change native path to 'lib'
Mono tries to prepend `lib` to the path when it doesn't exist,
without examining whether its a directory or a file.

So if our native libraries are in `NativeBinaries/linux/libgit2.so`,
it will try to load `libNativeBinaries/linux/libgit2.so`.  Changing
our directory name to `lib` will prevent this ridiculousness.